Extract method exyracyTaskAndDueDate from Class Deadline to Class Parser
