+++ data/output.txt
@@ -1,3 +1,5 @@
-T | 0 | homework
+T | 1 | homework
 T | 0 | collect milo
-E | 1 | clbv study | tomorrow | evening
+E | 0 | clbv study | tomorrow | evening
+D | 0 | ee2026 quiz | tomorrow
+D | 0 | club | tonight
