+++ src/main/java/omoh/tasktypes/Deadline.java
@@ -2,6 +2,7 @@ package omoh.tasktypes;
 
 import omoh.Omoh;
 
+import omoh.Parser;
 import omoh.customexceptions.IllegalDeadlineInput;
 
 public class Deadline extends Task {
@@ -19,23 +20,23 @@ public class Deadline extends Task {
         return "[D]" + super.toString() + " (by: " + by + ")";
     }
 
-    public static Deadline extractTaskAndDueDate (String input) throws IllegalDeadlineInput {
-        Deadline taskAndDeadlineString = new Deadline("random","random");
-        //splits string according to /by keyword
-        String[] parts = input.split("/by");
-        //if array is not size 2, means that we are missing inputs
-        if (parts.length != 2) {
-            throw new IllegalDeadlineInput();
-        }
-        //extracts task portion from input, after deadline keyword
-        taskAndDeadlineString.description = parts[0].substring("deadline".length()).trim();
-        //extracts deadline portion from input
-        taskAndDeadlineString.by = parts[1].trim();
-        if (taskAndDeadlineString.description.isEmpty() || taskAndDeadlineString.by.isEmpty()) {
-            throw new IllegalDeadlineInput();
-        }
-        return taskAndDeadlineString;
-    }
+//    public static Deadline extractTaskAndDueDate (String input) throws IllegalDeadlineInput {
+//        Deadline taskAndDeadlineString = new Deadline("random","random");
+//        //splits string according to /by keyword
+//        String[] parts = input.split("/by");
+//        //if array is not size 2, means that we are missing inputs
+//        if (parts.length != 2) {
+//            throw new IllegalDeadlineInput();
+//        }
+//        //extracts task portion from input, after deadline keyword
+//        taskAndDeadlineString.description = parts[0].substring("deadline".length()).trim();
+//        //extracts deadline portion from input
+//        taskAndDeadlineString.by = parts[1].trim();
+//        if (taskAndDeadlineString.description.isEmpty() || taskAndDeadlineString.by.isEmpty()) {
+//            throw new IllegalDeadlineInput();
+//        }
+//        return taskAndDeadlineString;
+//    }
 
     public String getBy() {
         return by;
@@ -52,7 +53,8 @@ public class Deadline extends Task {
 
     public static void addDeadline (String input) {
         try {
-            Deadline extractedInfo = Deadline.extractTaskAndDueDate(input);
+//            Deadline extractedInfo = Parser.extractTaskAndDueDate(input);
+            Deadline extractedInfo = Parser.extractTaskAndDueDate(input);
             Task.tasks.add(new Deadline(extractedInfo.description, extractedInfo.by));
             Task.totalTasks++;
             Deadline.printDeadline(extractedInfo);
